+++ src/app/Fake.DotNet.Paket/Paket.fs
@@ -147,6 +147,14 @@ let Pack setParams =
 let PushFiles setParams files =
     let parameters : PaketPushParams = PaketPushDefaults() |> setParams
 
+    TraceSecrets.register parameters.ApiKey "<PaketApiKey>"
+    match Environment.environVarOrNone "nugetkey" with
+    | Some k -> TraceSecrets.register k "<PaketApiKey>"
+    | None -> ()
+    match Environment.environVarOrNone "nuget-key" with
+    | Some k -> TraceSecrets.register k "<PaketApiKey>"
+    | None -> ()
+    
     let packages = Seq.toList files
     let url = if String.IsNullOrWhiteSpace parameters.PublishUrl then "" else " --url " + Process.toParam parameters.PublishUrl
     let endpoint = if String.IsNullOrWhiteSpace parameters.EndPoint then "" else " --endpoint " + Process.toParam parameters.EndPoint